test: migrate tests from test_prior.py to unit/test_prior.py
